server/requirements.txt
Flask==2.1.1
gunicorn==20.1.0
vertexai
pandas
Werkzeug==2.2.2
google-cloud-storage
google-cloud-aiplatform